Python/discord-bot/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# discord-bot/cogs/eval.py: 1,2
# discord-bot/main.py: 1,2
discord_ext_menus == 1.0.0a26+g84caae8

# discord-bot/cogs/eval.py: 8
import_expression == 1.1.4
